+++ CHANGELOG.rst
@@ -37,6 +37,7 @@ Motions:
  - Added multi select action to manage submitters, tags, states and
    recommendations [#4037, #4132].
  - Added timestampes for motions [#4134].
+ - New config option to set reason as required field [#4232]
 
 User:
  - Added new admin group which grants all permissions. Users of existing group